Dm android studio

+++ app/src/main/java/com/nlh/minishoping/Product.java
@@ -4,9 +4,9 @@ public class Product {
     public int ID;
     public String name;
     public String imageLink;
-    public double price;
+    public int price;
 
-    Product(int ID, String name, String imageLink, double price) {
+    Product(int ID, String name, String imageLink, int price) {
         this.ID = ID;
         this.name = name;
         this.imageLink = imageLink;
